requirements.txt
# google-api-python-client==2.128.0
# gspread==6.1.0
# moviepy==1.0.3
# python-dotenv==1.0.0
# schedule==1.2.1
# streamlit==1.30.0
# streamlit-authenticator==0.3.1
# oauth2client==4.1.3
# openai==1.25.2
# openpyxl==3.1.2
# extra-streamlit-components==0.1.71
# PyYAML==6.0.1
# pandas==1.5.3
# requests==2.31.0
# zipp==3.18.1
# google-api-core==2.19.0
# google-auth==2.29.0
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.0
# google-pasta==0.2.0
google-api-python-client==2.128.0
streamlit-chunk-file-uploader==1.0.4
streamlit-extras==0.4.3
gspread==6.1.0
moviepy==1.0.3
python-dotenv==1.0.0
schedule==1.2.1
streamlit==1.35.0
streamlit-authenticator==0.3.1
oauth2client==4.1.3
openai==1.25.2
openpyxl==3.1.2
extra-streamlit-components==0.1.71
imageio-ffmpeg==0.4.9